Extract DataCite XML parsing to service
Move DataCite XML extraction logic out of UploadXmlController into a dedicated parser service. Added DataCiteXmlImportParser and DataCiteXmlImportResult plus multiple SectionParser classes (Author, Contributor, Coverage, Date, Description, FundingReference, GcmdKeyword, Identifier, IsoContact, RelatedItem, RelatedWorkAndInstrument, Rights, Title) to encapsulate parsing responsibilities. UploadXmlController now delegates parsing to the import parser and stores the parser result via toSessionPayload(), removing a large amount of inline XML extraction helpers and logic. Updated tests (tests/pest/Arch/ArchTest.php) accordingly. This refactor improves separation of concerns, readability and testability of the XML import flow.
