+++ app/Http/Controllers/UploadXmlController.php
@@ -6,103 +6,20 @@ namespace App\Http\Controllers;
 
 use App\Enums\UploadErrorCode;
 use App\Http\Requests\UploadXmlRequest;
-use App\Models\ResourceType;
-use App\Services\RorLookupService;
 use App\Services\UploadLogService;
-use App\Support\GcmdUriHelper;
-use App\Support\MslLaboratoryService;
+use App\Services\Xml\DataCiteXmlImportParser;
 use App\Support\UploadError;
-use App\Support\UriHelper;
-use App\Support\Xml\XmlElementHelpers;
-use App\Support\XmlKeywordExtractor;
 use Illuminate\Http\JsonResponse;
-use Illuminate\Support\Facades\Log;
 use Illuminate\Support\Str;
-use Saloon\XmlWrangler\Data\Element;
 use Saloon\XmlWrangler\Exceptions\XmlReaderException;
 use Saloon\XmlWrangler\XmlReader;
 use VeeWee\Xml\Exception\RuntimeException as XmlRuntimeException;
 
 class UploadXmlController extends Controller
 {
-    /**
-     * @var array<string, string>
-     */
-    private const CONTRIBUTOR_ROLE_LABELS = [
-        'contactperson' => 'Contact Person',
-        'datacollector' => 'Data Collector',
-        'datacurator' => 'Data Curator',
-        'datamanager' => 'Data Manager',
-        'distributor' => 'Distributor',
-        'editor' => 'Editor',
-        'hostinginstitution' => 'Hosting Institution',
-        'producer' => 'Producer',
-        'projectleader' => 'Project Leader',
-        'projectmanager' => 'Project Manager',
-        'projectmember' => 'Project Member',
-        'registrationagency' => 'Registration Agency',
-        'registrationauthority' => 'Registration Authority',
-        'relatedperson' => 'Related Person',
-        'researcher' => 'Researcher',
-        'researchgroup' => 'Research Group',
-        'rightsholder' => 'Rights Holder',
-        'sponsor' => 'Sponsor',
-        'supervisor' => 'Supervisor',
-        'translator' => 'Translator',
-        'workpackageleader' => 'Work Package Leader',
-        'other' => 'Other',
-    ];
-
-    /**
-     * @var string[]
-     *
-     * @see resources/js/lib/contributors.ts INSTITUTION_ONLY_ROLE_KEY_VALUES
-     */
-    private const INSTITUTION_ONLY_CONTRIBUTOR_ROLE_KEYS = [
-        'distributor',
-        'hostinginstitution',
-        'registrationagency',
-        'registrationauthority',
-        'researchgroup',
-        'sponsor',
-    ];
-
-    /**
-     * Canonical DataCite related identifier types supported by ERNIE editor.
-     *
-     * @var string[]
-     */
-    private const RELATED_IDENTIFIER_TYPES = [
-        'DOI', 'URL', 'Handle', 'IGSN', 'URN', 'ISBN', 'ISSN', 'PURL', 'ARK',
-        'arXiv', 'bibcode', 'CSTR', 'EAN13', 'EISSN', 'ISTC', 'LISSN', 'LSID',
-        'PMID', 'RAiD', 'RRID', 'SWHID', 'UPC', 'w3id',
-    ];
-
-    /**
-     * Canonical DataCite relation types supported by ERNIE editor.
-     *
-     * @var string[]
-     */
-    private const RELATED_RELATION_TYPES = [
-        'Cites', 'IsCitedBy', 'References', 'IsReferencedBy',
-        'Documents', 'IsDocumentedBy', 'Describes', 'IsDescribedBy',
-        'IsNewVersionOf', 'IsPreviousVersionOf', 'HasVersion', 'IsVersionOf',
-        'HasTranslation', 'IsTranslationOf',
-        'Continues', 'IsContinuedBy', 'Obsoletes', 'IsObsoletedBy',
-        'IsVariantFormOf', 'IsOriginalFormOf', 'IsIdenticalTo',
-        'HasPart', 'IsPartOf', 'Compiles', 'IsCompiledBy',
-        'IsSourceOf', 'IsDerivedFrom',
-        'IsSupplementTo', 'IsSupplementedBy',
-        'Requires', 'IsRequiredBy',
-        'HasMetadata', 'IsMetadataFor',
-        'Reviews', 'IsReviewedBy',
-        'IsPublishedIn', 'Collects', 'IsCollectedBy',
-        'Other',
-    ];
-
     public function __construct(
         private readonly UploadLogService $uploadLogService,
-        private readonly RorLookupService $rorLookupService,
+        private readonly DataCiteXmlImportParser $importParser,
     ) {}
 
     public function __invoke(UploadXmlRequest $request): JsonResponse
@@ -124,105 +41,7 @@ class UploadXmlController extends Controller
 
         try {
             $reader = XmlReader::fromString($contents);
-
-            $doi = $this->extractFirstStringFromQuery(
-                $reader->xpathValue('//*[local-name()="identifier" and @identifierType="DOI"]'),
-            );
-            $year = $this->extractFirstStringFromQuery(
-                $reader->xpathValue('//*[local-name()="publicationYear"]'),
-            );
-            $version = $this->extractFirstStringFromQuery(
-                $reader->xpathValue('//*[local-name()="version"]'),
-            );
-            $language = $this->extractFirstStringFromQuery(
-                $reader->xpathValue('//*[local-name()="language"]'),
-            );
-            $authors = $this->extractAuthors($reader);
-            $contributorsAndLabs = $this->extractContributorsAndMslLaboratories($reader);
-            $contributors = $contributorsAndLabs['contributors'];
-            $mslLaboratories = $contributorsAndLabs['mslLaboratories'];
-            $contactPersons = $contributorsAndLabs['contactPersons'];
-
-            // Extract contact info (email, website) from ISO 19115 part of the XML
-            $isoContactInfo = $this->extractContactInfoFromIso($reader);
-
-            // Merge ContactPerson contributors into authors with isContact flag and contact info
-            $authors = $this->mergeContactPersonsIntoAuthors($authors, $contactPersons, $isoContactInfo);
-
-            $descriptions = $this->extractDescriptions($reader);
-            $dates = $this->extractDates($reader);
-            $coverages = $this->extractCoverages($reader, $dates);
-
-            // Remove Coverage dates from dates array (they are now in coverages)
-            // and strip the rawValue key used only for coverage time extraction
-            $dates = array_filter(
-                $dates,
-                fn (array $date) => ($date['dateType'] ?? '') !== 'coverage',
-            );
-            $dates = array_values(array_map(
-                fn (array $date) => array_diff_key($date, ['rawValue' => true]),
-                $dates,
-            ));
-
-            $gcmdKeywords = $this->extractGcmdKeywords($reader);
-
-            // Use dedicated service for keyword extraction
-            $keywordExtractor = new XmlKeywordExtractor;
-            $freeKeywords = $keywordExtractor->extractFreeKeywords($reader);
-            $mslKeywords = $keywordExtractor->extractMslKeywords($reader);
-            $gemetKeywords = $keywordExtractor->extractGemetKeywords($reader);
-
-            $rightsElements = $reader
-                ->xpathElement('//*[local-name()="rightsList"]/*[local-name()="rights"]')
-                ->get();
-            $licenses = [];
-
-            foreach ($rightsElements as $element) {
-                $identifier = $element->getAttribute('rightsIdentifier');
-                if ($identifier) {
-                    $licenses[] = $identifier;
-                }
-            }
-
-            $titleElements = $reader
-                ->xpathElement('//*[local-name()="resource"]/*[local-name()="titles"]/*[local-name()="title"]')
-                ->get();
-            $titles = [];
-
-            foreach ($titleElements as $element) {
-                $titleType = $element->getAttribute('titleType');
-                $xmlLang = $element->getAttribute('xml:lang');
-                $titles[] = [
-                    'title' => $element->getContent(),
-                    'titleType' => $titleType ? Str::kebab($titleType) : 'main-title',
-                    'language' => is_string($xmlLang) && trim($xmlLang) !== '' ? trim($xmlLang) : null,
-                ];
-            }
-
-            $mainTitles = array_values(array_filter(
-                $titles,
-                fn ($t) => $t['titleType'] === 'main-title'
-            ));
-            $otherTitles = array_values(array_filter(
-                $titles,
-                fn ($t) => $t['titleType'] !== 'main-title'
-            ));
-            $titles = array_merge($mainTitles, $otherTitles);
-
-            $resourceTypeElement = $this->extractFirstElementFromQuery(
-                $reader->xpathElement('//*[local-name()="resourceType"]'),
-            );
-            $resourceTypeName = $resourceTypeElement?->getAttribute('resourceTypeGeneral');
-            $resourceType = null;
-
-            if ($resourceTypeName !== null) {
-                $resourceTypeModel = ResourceType::whereRaw('LOWER(name) = ?', [Str::lower($resourceTypeName)])->first();
-                $resourceType = $resourceTypeModel?->id;
-            }
-
-            ['relatedWorks' => $relatedWorks, 'instruments' => $instruments] = $this->extractRelatedWorksAndInstruments($reader, $filename);
-            $fundingReferences = $this->extractFundingReferences($reader);
-            $relatedItems = $this->extractRelatedItems($reader);
+            $result = $this->importParser->parse($reader, $filename);
         } catch (XmlReaderException|XmlRuntimeException $e) {
             $error = UploadError::withMessage(
                 UploadErrorCode::XML_PARSE_ERROR,
@@ -256,2325 +75,11 @@ class UploadXmlController extends Controller
 
         // Store data in session to avoid 414 URI Too Long errors
         $sessionKey = 'xml_upload_'.Str::random(32);
-
-        session()->put($sessionKey, [
-            'doi' => $doi,
-            'year' => $year,
-            'version' => $version,
-            'language' => $language,
-            'resourceType' => $resourceType !== null ? (string) $resourceType : null,
-            'titles' => $titles,
-            'licenses' => $licenses,
-            'authors' => $authors,
-            'contributors' => $contributors,
-            'descriptions' => $descriptions,
-            'dates' => $dates,
-            'coverages' => $coverages,
-            'relatedWorks' => $relatedWorks,
-            'instruments' => $instruments,
-            'gcmdKeywords' => $gcmdKeywords,
-            'freeKeywords' => $freeKeywords,
-            'mslKeywords' => $mslKeywords,
-            'gemetKeywords' => $gemetKeywords,
-            'fundingReferences' => $fundingReferences,
-            'mslLaboratories' => $mslLaboratories,
-            'relatedItems' => $relatedItems,
-        ]);
+        session()->put($sessionKey, $result->toSessionPayload());
 
         return response()->json([
             'sessionKey' => $sessionKey,
         ]);
-    }
-
-    private function extractFirstStringFromQuery(mixed $query): ?string
-    {
-        return XmlElementHelpers::firstStringFromQuery($query);
-    }
-
-    private function extractFirstElementFromQuery(mixed $query): ?Element
-    {
-        return XmlElementHelpers::firstElementFromQuery($query);
-    }
-
-    /**
-     * @return array<int, array<string, mixed>>
-     */
-    private function extractAuthors(XmlReader $reader): array
-    {
-        $creatorElements = $reader
-            ->xpathElement('//*[local-name()="resource"]/*[local-name()="creators"]/*[local-name()="creator"]')
-            ->get();
-
-        $authors = [];
-
-        foreach ($creatorElements as $creator) {
-            $content = $creator->getContent();
-
-            if (! is_array($content)) {
-                continue;
-            }
-
-            $creatorName = $this->firstElement($content, 'creatorName');
-            $nameType = $creatorName?->getAttribute('nameType');
-            $type = is_string($nameType) && Str::lower($nameType) === 'organizational' ? 'institution' : 'person';
-
-            $affiliations = $this->extractAffiliations($content);
-
-            if ($type === 'institution') {
-                $authors[] = [
-                    'type' => 'institution',
-                    'institutionName' => $this->stringValue($creatorName) ?? '',
-                    'affiliations' => $affiliations,
-                ];
-
-                continue;
-            }
-
-            $givenName = $this->stringValue($this->firstElement($content, 'givenName'));
-            $familyName = $this->stringValue($this->firstElement($content, 'familyName'));
-
-            if ((! $givenName || ! $familyName) && $creatorName instanceof Element) {
-                $resolved = $this->splitCreatorName($this->stringValue($creatorName));
-                $familyName ??= $resolved['familyName'];
-                $givenName ??= $resolved['givenName'];
-            }
-
-            $authors[] = [
-                'type' => 'person',
-                'orcid' => $this->extractOrcid($content),
-                'firstName' => $givenName ?? '',
-                'lastName' => $familyName ?? ($this->stringValue($creatorName) ?? ''),
-                'affiliations' => $affiliations,
-            ];
-        }
-
-        return $authors;
-    }
-
-    /**
-     * Extract related identifiers from DataCite XML for editor related works prefill.
-     *
-     * Instrument PIDs (relationType="IsCollectedBy" with identifierType="Handle")
-     * are separated into their own array for the "Used Instruments" form section.
-     *
-     * @return array{relatedWorks: array<int, array{identifier: string, identifier_type: string, relation_type: string, relation_type_information: string|null, position: int}>, instruments: array<int, array{pid: string, pidType: string, name: string}>}
-     */
-    private function extractRelatedWorksAndInstruments(XmlReader $reader, string $filename): array
-    {
-        $relatedIdentifierElements = $reader
-            ->xpathElement('//*[local-name()="resource"]/*[local-name()="relatedIdentifiers"]/*[local-name()="relatedIdentifier"]')
-            ->get();
-
-        /** @var array<string, string> $identifierTypeLookup */
-        $identifierTypeLookup = [];
-        foreach (self::RELATED_IDENTIFIER_TYPES as $name) {
-            $identifierTypeLookup[mb_strtolower($name)] = $name;
-        }
-
-        /** @var array<string, string> $relationTypeLookup */
-        $relationTypeLookup = [];
-        foreach (self::RELATED_RELATION_TYPES as $name) {
-            $relationTypeLookup[mb_strtolower($name)] = $name;
-        }
-
-        $relatedWorks = [];
-        $instruments = [];
-
-        foreach ($relatedIdentifierElements as $index => $element) {
-            $identifier = $this->stringValue($element);
-            $identifierTypeRaw = $element->getAttribute('relatedIdentifierType');
-            $relationTypeRaw = $element->getAttribute('relationType');
-            $relationTypeInformationRaw = $element->getAttribute('relationTypeInformation');
-
-            if (! is_string($identifier) || $identifier === '') {
-                continue;
-            }
-
-            $identifierType = is_string($identifierTypeRaw)
-                ? ($identifierTypeLookup[mb_strtolower(trim($identifierTypeRaw))] ?? null)
-                : null;
-            $relationType = is_string($relationTypeRaw)
-                ? ($relationTypeLookup[mb_strtolower(trim($relationTypeRaw))] ?? null)
-                : null;
-
-            if ($identifierType === null || $relationType === null) {
-                Log::warning('Skipping related identifier with unsupported type values during XML upload', [
-                    'filename' => $filename,
-                    'index' => $index,
-                    'identifier' => $identifier,
-                    'relatedIdentifierType' => $identifierTypeRaw,
-                    'relationType' => $relationTypeRaw,
-                ]);
-
-                continue;
-            }
-
-            // Separate instrument PIDs (IsCollectedBy + Handle) into their own array
-            if ($relationType === 'IsCollectedBy' && $identifierType === 'Handle') {
-                $instruments[] = [
-                    'pid' => $identifier,
-                    'pidType' => $identifierType,
-                    'name' => $identifier, // PID as fallback; UsedInstrumentsField resolves the name from vocabulary on load
-                ];
-
-                continue;
-            }
-
-            $relatedWorks[] = [
-                'identifier' => $identifier,
-                'identifier_type' => $identifierType,
-                'relation_type' => $relationType,
-                'relation_type_information' => is_string($relationTypeInformationRaw) && trim($relationTypeInformationRaw) !== '' ? trim($relationTypeInformationRaw) : null,
-                'position' => count($relatedWorks),
-            ];
-        }
-
-        return [
-            'relatedWorks' => $relatedWorks,
-            'instruments' => $instruments,
-        ];
-    }
-
-    /**
-     * Extract `<relatedItems>` (DataCite 4.7) with nested titles, creators,
-     * contributors and affiliations into a form payload.
-     *
-     * The returned structure mirrors the JSON accepted by the REST endpoints
-     * `POST /resources/{resource}/related-items` so it can later be persisted
-     * via `RelatedItemStorageService`.
-     *
-     * @return array<int, array<string, mixed>>
-     */
-    private function extractRelatedItems(XmlReader $reader): array
-    {
-        $items = $reader
-            ->xpathElement('//*[local-name()="resource"]/*[local-name()="relatedItems"]/*[local-name()="relatedItem"]')
-            ->get();
-
-        $result = [];
-
-        foreach ($items as $position => $item) {
-            $relatedItemType = $item->getAttribute('relatedItemType');
-            $relationTypeSlug = $item->getAttribute('relationType');
-
-            if (! is_string($relatedItemType) || trim($relatedItemType) === '') {
-                continue;
-            }
-            if (! is_string($relationTypeSlug) || trim($relationTypeSlug) === '') {
-                continue;
-            }
-
-            $titles = $this->extractRelatedItemTitles($item);
-            if ($titles === []) {
-                // DataCite 4.7 requires at least one title with type MainTitle.
-                continue;
-            }
-
-            $entry = [
-                'related_item_type' => trim($relatedItemType),
-                'relation_type_slug' => trim($relationTypeSlug),
-                'titles' => $titles,
-                'creators' => $this->extractRelatedItemCreators($item),
-                'contributors' => $this->extractRelatedItemContributors($item),
-                'publication_year' => $this->intOrNull($this->scalarChild($item, 'publicationYear')),
-                'volume' => $this->stringOrNull($this->scalarChild($item, 'volume')),
-                'issue' => $this->stringOrNull($this->scalarChild($item, 'issue')),
-                'first_page' => $this->stringOrNull($this->scalarChild($item, 'firstPage')),
-                'last_page' => $this->stringOrNull($this->scalarChild($item, 'lastPage')),
-                'publisher' => $this->stringOrNull($this->scalarChild($item, 'publisher')),
-                'edition' => $this->stringOrNull($this->scalarChild($item, 'edition')),
-                'position' => $position,
-            ];
-
-            // <number numberType="Article">42</number>
-            $numberElement = $this->firstChildElement($item, 'number');
-            if ($numberElement !== null) {
-                $numberValue = $this->stringValue($numberElement);
-                if (is_string($numberValue) && trim($numberValue) !== '') {
-                    $entry['number'] = trim($numberValue);
-                    $typeAttr = $numberElement->getAttribute('numberType');
-                    $entry['number_type'] = is_string($typeAttr) && trim($typeAttr) !== '' ? trim($typeAttr) : null;
-                }
-            }
-
-            // <relatedItemIdentifier relatedItemIdentifierType="DOI">10.1234/x</relatedItemIdentifier>
-            $identifierElement = $this->firstChildElement($item, 'relatedItemIdentifier');
-            if ($identifierElement !== null) {
-                $identifierValue = $this->stringValue($identifierElement);
-                if (is_string($identifierValue) && trim($identifierValue) !== '') {
-                    $entry['identifier'] = trim($identifierValue);
-                    $idType = $identifierElement->getAttribute('relatedItemIdentifierType');
-                    $entry['identifier_type'] = is_string($idType) && trim($idType) !== '' ? trim($idType) : null;
-                    $scheme = $identifierElement->getAttribute('relatedMetadataScheme');
-                    if (is_string($scheme) && trim($scheme) !== '') {
-                        $entry['related_metadata_scheme'] = trim($scheme);
-                    }
-                    $schemeUri = $identifierElement->getAttribute('schemeURI');
-                    if (is_string($schemeUri) && trim($schemeUri) !== '') {
-                        $entry['scheme_uri'] = trim($schemeUri);
-                    }
-                    $schemeType = $identifierElement->getAttribute('schemeType');
-                    if (is_string($schemeType) && trim($schemeType) !== '') {
-                        $entry['scheme_type'] = trim($schemeType);
-                    }
-                }
-            }
-
-            $result[] = $entry;
-        }
-
-        return $result;
-    }
-
-    /**
-     * @return array<int, array{title: string, title_type: string}>
-     */
-    private function extractRelatedItemTitles(Element $item): array
-    {
-        $titles = [];
-        foreach ($this->childElements($item, 'titles') as $titlesWrapper) {
-            foreach ($this->childElements($titlesWrapper, 'title') as $t) {
-                $value = $this->stringValue($t);
-                if (! is_string($value) || trim($value) === '') {
-                    continue;
-                }
-                $typeAttr = $t->getAttribute('titleType');
-                $titleType = is_string($typeAttr) && trim($typeAttr) !== '' ? trim($typeAttr) : 'MainTitle';
-                $titles[] = [
-                    'title' => trim($value),
-                    'title_type' => $titleType,
-                ];
-            }
-        }
-
-        $hasMain = false;
-        foreach ($titles as $t) {
-            if ($t['title_type'] === 'MainTitle') {
-                $hasMain = true;
-
-                break;
-            }
-        }
-        if (! $hasMain && $titles !== []) {
-            // Promote the first entry to MainTitle so the item passes validation.
-            $titles[0]['title_type'] = 'MainTitle';
-        }
-
-        return $titles;
-    }
-
-    /**
-     * @return array<int, array<string, mixed>>
-     */
-    private function extractRelatedItemCreators(Element $item): array
-    {
-        $result = [];
-        foreach ($this->childElements($item, 'creators') as $creatorsEl) {
-            foreach ($this->childElements($creatorsEl, 'creator') as $creator) {
-                $entry = $this->extractRelatedItemPerson($creator, 'creatorName');
-                if ($entry !== null) {
-                    $result[] = $entry;
-                }
-            }
-        }
-
-        return $result;
-    }
-
-    /**
-     * @return array<int, array<string, mixed>>
-     */
-    private function extractRelatedItemContributors(Element $item): array
-    {
-        $result = [];
-        foreach ($this->childElements($item, 'contributors') as $contributorsEl) {
-            foreach ($this->childElements($contributorsEl, 'contributor') as $contributor) {
-                $entry = $this->extractRelatedItemPerson($contributor, 'contributorName');
-                if ($entry === null) {
-                    continue;
-                }
-                $type = $contributor->getAttribute('contributorType');
-                if (is_string($type) && trim($type) !== '') {
-                    $entry['contributor_type'] = trim($type);
-                    $result[] = $entry;
-                }
-            }
-        }
-
-        return $result;
-    }
-
-    /**
-     * @return array<string, mixed>|null
-     */
-    private function extractRelatedItemPerson(Element $person, string $nameElementName): ?array
-    {
-        $nameElement = $this->firstChildElement($person, $nameElementName);
-        if ($nameElement === null) {
-            return null;
-        }
-
-        $nameValue = $this->stringValue($nameElement);
-        if (! is_string($nameValue) || trim($nameValue) === '') {
-            return null;
-        }
-
-        $nameTypeAttr = $nameElement->getAttribute('nameType');
-        $nameType = is_string($nameTypeAttr) && trim($nameTypeAttr) !== '' ? trim($nameTypeAttr) : 'Personal';
-
-        $entry = [
-            'name_type' => $nameType,
-            'name' => trim($nameValue),
-            'given_name' => $this->stringOrNull($this->scalarChild($person, 'givenName')),
-            'family_name' => $this->stringOrNull($this->scalarChild($person, 'familyName')),
-        ];
-
-        $idElement = $this->firstChildElement($person, 'nameIdentifier');
-        if ($idElement !== null) {
-            $idValue = $this->stringValue($idElement);
-            if (is_string($idValue) && trim($idValue) !== '') {
-                $entry['name_identifier'] = trim($idValue);
-                $scheme = $idElement->getAttribute('nameIdentifierScheme');
-                if (is_string($scheme) && trim($scheme) !== '') {
-                    $entry['name_identifier_scheme'] = trim($scheme);
-                }
-            }
-        }
-
-        $affiliations = [];
-        foreach ($this->childElements($person, 'affiliation') as $aff) {
-            $affValue = $this->stringValue($aff);
-            if (! is_string($affValue) || trim($affValue) === '') {
-                continue;
-            }
-            $affEntry = ['name' => trim($affValue)];
-            $affId = $aff->getAttribute('affiliationIdentifier');
-            if (is_string($affId) && trim($affId) !== '') {
-                $affEntry['affiliation_identifier'] = trim($affId);
-            }
-            $affScheme = $aff->getAttribute('affiliationIdentifierScheme');
-            if (is_string($affScheme) && trim($affScheme) !== '') {
-                $affEntry['scheme'] = trim($affScheme);
-            }
-            $affiliations[] = $affEntry;
-        }
-        if ($affiliations !== []) {
-            $entry['affiliations'] = $affiliations;
-        }
-
-        return $entry;
-    }
-
-    /**
-     * @return Element[]
-     */
-    private function childElements(Element $parent, string $localName): array
-    {
-        return XmlElementHelpers::childElements($parent, $localName);
-    }
-
-    private function firstChildElement(Element $parent, string $localName): ?Element
-    {
-        return XmlElementHelpers::firstChildElement($parent, $localName);
-    }
-
-    private function scalarChild(Element $parent, string $localName): ?string
-    {
-        return XmlElementHelpers::scalarChild($parent, $localName);
-    }
-
-    private function stringOrNull(?string $value): ?string
-    {
-        return XmlElementHelpers::stringOrNull($value);
-    }
-
-    private function intOrNull(?string $value): ?int
-    {
-        return XmlElementHelpers::intOrNull($value);
-    }
-
-    /**
-     * Extract both regular contributors and MSL laboratories
-     *
-     * @return array{contributors: array<int, array<string, mixed>>, mslLaboratories: array<int, array<string, string>>, contactPersons: array<int, array<string, mixed>>}
-     */
-    private function extractContributorsAndMslLaboratories(XmlReader $reader): array
-    {
-        $contributorElements = $reader
-            ->xpathElement('//*[local-name()="resource"]/*[local-name()="contributors"]/*[local-name()="contributor"]')
-            ->get();
-
-        $contributors = [];
-        $contributorIndexByKey = [];
-        $mslLaboratories = [];
-        $contactPersons = [];
-
-        foreach ($contributorElements as $contributor) {
-            $content = $contributor->getContent();
-
-            if (! is_array($content)) {
-                continue;
-            }
-
-            $contributorType = $contributor->getAttribute('contributorType');
-            $roles = $this->extractContributorRoles($contributorType);
-
-            // Check if this is a ContactPerson - handle separately
-            if (strcasecmp($contributorType ?? '', 'ContactPerson') === 0) {
-                $contactPerson = $this->extractContactPersonData($content);
-                if ($contactPerson !== null) {
-                    $contactPersons[] = $contactPerson;
-                }
-
-                continue; // Don't add to regular contributors
-            }
-
-            // Check if this is an MSL Laboratory (HostingInstitution with labid)
-            // Use contributorType directly for robust detection (case-insensitive)
-            $labId = $this->extractLabId($content);
-            if ($labId !== null && strcasecmp($contributorType ?? '', 'HostingInstitution') === 0) {
-                // This is an MSL Laboratory
-                Log::info('Extracting MSL Laboratory', [
-                    'labId' => $labId,
-                    'contributorType' => $contributorType,
-                ]);
-
-                $mslLab = $this->extractMslLaboratory($content, $labId);
-
-                if ($mslLab !== null) {
-                    Log::info('MSL Laboratory extracted successfully', $mslLab);
-                    $mslLaboratories[] = $mslLab;
-                } else {
-                    Log::warning('MSL Laboratory extraction returned null', [
-                        'labId' => $labId,
-                    ]);
-                }
-
-                continue; // Don't add to contributors
-            }
-
-            $nameElement = $this->firstElement($content, 'contributorName');
-            $nameType = $nameElement?->getAttribute('nameType');
-
-            $isInstitution = $this->isInstitutionContributor($nameType, $roles);
-
-            if ($isInstitution) {
-                $institutionName = $this->stringValue($nameElement) ?? '';
-                $affiliations = $this->extractInstitutionContributorAffiliations($content, $institutionName);
-
-                $contributorData = [
-                    'type' => 'institution',
-                    'institutionName' => $institutionName,
-                    'roles' => $roles,
-                    'affiliations' => $affiliations,
-                ];
-
-                $contributors = $this->storeContributor(
-                    $contributors,
-                    $contributorIndexByKey,
-                    $contributorData,
-                );
-
-                continue;
-            }
-
-            $givenName = $this->stringValue($this->firstElement($content, 'givenName'));
-            $familyName = $this->stringValue($this->firstElement($content, 'familyName'));
-
-            if ((! $givenName || ! $familyName) && $nameElement instanceof Element) {
-                $resolved = $this->splitCreatorName($this->stringValue($nameElement));
-                $familyName ??= $resolved['familyName'];
-                $givenName ??= $resolved['givenName'];
-            }
-
-            $fallbackLastName = $familyName ?? ($this->stringValue($nameElement) ?? '');
-
-            $contributorData = [
-                'type' => 'person',
-                'roles' => $roles,
-                'orcid' => $this->extractOrcid($content),
-                'firstName' => $givenName ?? '',
-                'lastName' => $fallbackLastName,
-                'affiliations' => $this->extractAffiliations($content),
-            ];
-
-            $contributors = $this->storeContributor(
-                $contributors,
-                $contributorIndexByKey,
-                $contributorData,
-            );
-        }
-
-        return [
-            'contributors' => $contributors,
-            'mslLaboratories' => $mslLaboratories,
-            'contactPersons' => $contactPersons,
-        ];
-    }
-
-    /**
-     * Extract data from a ContactPerson contributor element.
-     *
-     * @param  array<string, mixed>  $content  Contributor element content
-     * @return array<string, mixed>|null  Contact person data or null if invalid
-     */
-    private function extractContactPersonData(array $content): ?array
-    {
-        $nameElement = $this->firstElement($content, 'contributorName');
-        $nameType = $nameElement?->getAttribute('nameType');
-
-        // Check if this is a person (not institution)
-        if (is_string($nameType) && Str::lower($nameType) === 'organizational') {
-            // Skip organizational contact persons (not supported in ERNIE author contact)
-            return null;
-        }
-
-        $givenName = $this->stringValue($this->firstElement($content, 'givenName'));
-        $familyName = $this->stringValue($this->firstElement($content, 'familyName'));
-
-        // Try to parse name from contributorName if separate parts are missing
-        if ((! $givenName || ! $familyName) && $nameElement instanceof Element) {
-            $resolved = $this->splitCreatorName($this->stringValue($nameElement));
-            $familyName ??= $resolved['familyName'];
-            $givenName ??= $resolved['givenName'];
-        }
-
-        // Require at least a family name
-        if (! $familyName) {
-            return null;
-        }
-
-        return [
-            'type' => 'person',
-            'orcid' => $this->extractOrcid($content),
-            'firstName' => $givenName ?? '',
-            'lastName' => $familyName,
-            'affiliations' => $this->extractAffiliations($content),
-        ];
-    }
-
-    /**
-     * @return array<int, array<string, string|null>>
-     */
-    private function extractDescriptions(XmlReader $reader): array
-    {
-        $descriptionElements = $reader
-            ->xpathElement('//*[local-name()="resource"]/*[local-name()="descriptions"]/*[local-name()="description"]')
-            ->get();
-
-        $descriptions = [];
-
-        foreach ($descriptionElements as $element) {
-            $descriptionType = $element->getAttribute('descriptionType');
-            $description = $element->getContent();
-
-            if (! is_string($description) || trim($description) === '') {
-                continue;
-            }
-
-            $descLang = $element->getAttribute('xml:lang');
-
-            $descriptions[] = [
-                'type' => $descriptionType ?? 'Other',
-                'description' => $description,
-                'language' => is_string($descLang) && trim($descLang) !== '' ? trim($descLang) : null,
-            ];
-        }
-
-        return $descriptions;
-    }
-
-    /**
-     * Normalize a date string to YYYY-MM-DD format.
-     *
-     * Handles various input formats:
-     * - Full date: "2024-01-15" -> "2024-01-15"
-     * - Year only: "2024" -> "2024-01-01"
-     * - Year-month: "2024-06" -> "2024-06-01"
-     * - DateTime: "2024-01-15 10:30:00" -> "2024-01-15"
-     * - Invalid/empty: returns empty string
-     *
-     * Note: Year-only dates default to January 1st, and year-month dates default to the 1st
-     * of the month. This is a pragmatic choice for date range comparisons and validation,
-     * but may not accurately represent the original semantic intent (e.g., "2024" meaning
-     * "sometime in 2024"). The original date value should be preserved separately if the
-     * full semantic meaning is important.
-     */
-    private function normalizeDateString(string $dateValue): string
-    {
-        $dateValue = trim($dateValue);
-
-        if ($dateValue === '') {
-            return '';
-        }
-
-        // Handle datetime format (strip time part): "2024-01-15 10:30:00" -> "2024-01-15"
-        if (str_contains($dateValue, ' ')) {
-            $dateValue = explode(' ', $dateValue)[0];
-        }
-
-        // Handle ISO datetime with T separator: "2024-01-15T10:30:00" -> "2024-01-15"
-        if (str_contains($dateValue, 'T')) {
-            $dateValue = explode('T', $dateValue)[0];
-        }
-
-        // Check if it's already a valid YYYY-MM-DD format
-        if (preg_match('/^\d{4}-\d{2}-\d{2}$/', $dateValue)) {
-            return $dateValue;
-        }
-
-        // Handle year-month format: "2024-06" -> "2024-06-01" (defaults to 1st of month)
-        if (preg_match('/^(\d{4})-(\d{2})$/', $dateValue, $matches)) {
-            return $matches[1].'-'.$matches[2].'-01';
-        }
-
-        // Handle year-only format: "2024" -> "2024-01-01" (defaults to January 1st)
-        if (preg_match('/^\d{4}$/', $dateValue)) {
-            return $dateValue.'-01-01';
-        }
-
-        // Try to parse with strtotime as last resort
-        // Warning: strtotime can produce unexpected results for ambiguous strings
-        $timestamp = strtotime($dateValue);
-        if ($timestamp !== false) {
-            Log::debug('Date normalization used strtotime fallback', [
-                'original' => $dateValue,
-                'normalized' => date('Y-m-d', $timestamp),
-            ]);
-
-            return date('Y-m-d', $timestamp);
-        }
-
-        Log::warning('Could not parse date value during XML import', [
-            'value' => $dateValue,
-        ]);
-
-        // Return empty string for unparseable dates
-        return '';
-    }
-
-    /**
-     * @return array<int, array<string, string>>
-     */
-    private function extractDates(XmlReader $reader): array
-    {
-        $dateElements = $reader
-            ->xpathElement('//*[local-name()="resource"]/*[local-name()="dates"]/*[local-name()="date"]')
-            ->get();
-
-        $dates = [];
-
-        foreach ($dateElements as $element) {
-            $dateType = $element->getAttribute('dateType');
-            $dateValue = $element->getContent();
-
-            if (! is_string($dateValue) || trim($dateValue) === '') {
-                continue;
-            }
-
-            $dateValue = trim($dateValue);
-
-            // Parse date value - can be single date or range (start/end)
-            $startDate = '';
-            $endDate = '';
-
-            if (str_contains($dateValue, '/')) {
-                // Date range format: "2024-01-01/2024-12-31" or "2010/2020" or open range "/2024-12-31"
-                [$start, $end] = explode('/', $dateValue, 2);
-                $startDate = $this->normalizeDateString(trim($start));
-                $endDate = $this->normalizeDateString(trim($end));
-            } else {
-                // Single date format: "2024-01-01" or "2024"
-                $startDate = $this->normalizeDateString($dateValue);
-            }
-
-            $dates[] = [
-                'dateType' => Str::kebab($dateType ?? 'other'),
-                'startDate' => $startDate,
-                'endDate' => $endDate,
-                'rawValue' => $dateValue,
-            ];
-        }
-
-        return $dates;
-    }
-
-    /**
-     * Extract spatial and temporal coverages from DataCite XML.
-     *
-     * DataCite stores:
-     * - Spatial coverage in <geoLocations> (with geoLocationPoint and geoLocationBox)
-     * - Temporal coverage as <date dateType="Coverage">
-     *
-     * @param  array<int, array<string, string>>  $dates  Already extracted dates (to find Coverage date)
-     * @return array<int, array<string, string|array<int, array<string, float>>>>
-     */
-    private function extractCoverages(XmlReader $reader, array $dates): array
-    {
-        $coverages = [];
-
-        // Extract temporal coverage from dates with type "coverage"
-        $temporalCoverage = null;
-        foreach ($dates as $date) {
-            if (($date['dateType'] ?? '') === 'coverage') {
-                $temporalCoverage = $date;
-                break;
-            }
-        }
-
-        // Parse temporal coverage with full datetime components (time + timezone)
-        $temporalComponents = $this->parseTemporalCoverageComponents($temporalCoverage);
-
-        // Extract geoLocations
-        $geoLocationElements = $reader
-            ->xpathElement('//*[local-name()="resource"]/*[local-name()="geoLocations"]/*[local-name()="geoLocation"]')
-            ->get();
-
-        if (count($geoLocationElements) === 0 && $temporalCoverage !== null) {
-            // Only temporal coverage, no spatial data
-            $coverages[] = [
-                'id' => 'coverage-1',
-                'type' => 'point',
-                'latMin' => '',
-                'latMax' => '',
-                'lonMin' => '',
-                'lonMax' => '',
-                'polygonPoints' => [],
-                'startDate' => $temporalComponents['startDate'],
-                'endDate' => $temporalComponents['endDate'],
-                'startTime' => $temporalComponents['startTime'],
-                'endTime' => $temporalComponents['endTime'],
-                'timezone' => $temporalComponents['timezone'],
-                'description' => '',
-            ];
-
-            return $coverages;
-        }
-
-        $index = 1;
-        foreach ($geoLocationElements as $geoLocationIndex => $geoLocation) {
-            $coverage = [
-                'id' => 'coverage-'.$index,
-                'type' => 'point',
-                'latMin' => '',
-                'latMax' => '',
-                'lonMin' => '',
-                'lonMax' => '',
-                'polygonPoints' => [],
-                'startDate' => $temporalComponents['startDate'],
-                'endDate' => $temporalComponents['endDate'],
-                'startTime' => $temporalComponents['startTime'],
-                'endTime' => $temporalComponents['endTime'],
-                'timezone' => $temporalComponents['timezone'],
-                'description' => '',
-            ];
-
-            // Build XPath base for this specific geoLocation element
-            // Use //* to find resource anywhere in document (supports envelope-wrapped XML)
-            // Cast to int to ensure type safety for PHPStan
-            $geoLocationPath = '//*[local-name()="resource"]/*[local-name()="geoLocations"]/*[local-name()="geoLocation"]['.((int) $geoLocationIndex + 1).']';
-
-            // Extract geoLocationPlace (description)
-            $place = $this->extractFirstStringFromQuery(
-                $reader->xpathValue($geoLocationPath.'/*[local-name()="geoLocationPlace"]')
-            );
-            if ($place !== null) {
-                $coverage['description'] = trim($place);
-            }
-
-            // Extract geoLocationPoint
-            $latText = $this->extractFirstStringFromQuery(
-                $reader->xpathValue($geoLocationPath.'/*[local-name()="geoLocationPoint"]/*[local-name()="pointLatitude"]')
-            );
-            $lonText = $this->extractFirstStringFromQuery(
-                $reader->xpathValue($geoLocationPath.'/*[local-name()="geoLocationPoint"]/*[local-name()="pointLongitude"]')
-            );
-
-            if ($latText !== null && $lonText !== null) {
-                $coverage['latMin'] = $this->formatCoordinate($latText);
-                $coverage['lonMin'] = $this->formatCoordinate($lonText);
-                // For points, leave latMax and lonMax empty (ERNIE convention)
-            }
-
-            // Extract geoLocationBox (takes precedence if both exist)
-            $west = $this->extractFirstStringFromQuery(
-                $reader->xpathValue($geoLocationPath.'/*[local-name()="geoLocationBox"]/*[local-name()="westBoundLongitude"]')
-            );
-            $east = $this->extractFirstStringFromQuery(
-                $reader->xpathValue($geoLocationPath.'/*[local-name()="geoLocationBox"]/*[local-name()="eastBoundLongitude"]')
-            );
-            $south = $this->extractFirstStringFromQuery(
-                $reader->xpathValue($geoLocationPath.'/*[local-name()="geoLocationBox"]/*[local-name()="southBoundLatitude"]')
-            );
-            $north = $this->extractFirstStringFromQuery(
-                $reader->xpathValue($geoLocationPath.'/*[local-name()="geoLocationBox"]/*[local-name()="northBoundLatitude"]')
-            );
-
-            if ($west !== null && $east !== null && $south !== null && $north !== null) {
-                $coverage['lonMin'] = $this->formatCoordinate($west);
-                $coverage['lonMax'] = $this->formatCoordinate($east);
-                $coverage['latMin'] = $this->formatCoordinate($south);
-                $coverage['latMax'] = $this->formatCoordinate($north);
-            }
-
-            // Extract geoLocationPolygon (highest precedence)
-            $polygonPoints = $this->extractPolygonPoints($reader, $geoLocationPath);
-            if (count($polygonPoints) > 0) {
-                $coverage['polygonPoints'] = $polygonPoints;
-            }
-
-            // Determine coverage type based on what was extracted
-            $coverage['type'] = $this->determineCoverageType($coverage);
-
-            // Only add coverage if it has at least coordinates or description
-            if ($coverage['latMin'] !== '' || $coverage['lonMin'] !== '' ||
-                ! empty($coverage['polygonPoints']) ||
-                $coverage['description'] !== '' || $coverage['startDate'] !== '') {
-                $coverages[] = $coverage;
-                $index++;
-            }
-        }
-
-        return $coverages;
-    }
-
-    /**
-     * Format coordinate value to max 6 decimal places.
-     */
-    private function formatCoordinate(string $value): string
-    {
-        $trimmed = trim($value);
-        if ($trimmed === '') {
-            return '';
-        }
-
-        $float = (float) $trimmed;
-
-        return number_format($float, 6, '.', '');
-    }
-
-    /**
-     * Extract polygon points from geoLocationPolygon element.
-     *
-     * DataCite format:
-     * <geoLocationPolygon>
-     *   <polygonPoint>
-     *     <pointLatitude>52.5</pointLatitude>
-     *     <pointLongitude>13.4</pointLongitude>
-     *   </polygonPoint>
-     *   ...
-     * </geoLocationPolygon>
-     *
-     * @return array<int, array{lat: float, lon: float}>
-     */
-    private function extractPolygonPoints(XmlReader $reader, string $geoLocationPath): array
-    {
-        $points = [];
-
-        // Get all polygonPoint elements within this geoLocationPolygon
-        $polygonPointElements = $reader
-            ->xpathElement($geoLocationPath.'/*[local-name()="geoLocationPolygon"]/*[local-name()="polygonPoint"]')
-            ->get();
-
-        foreach ($polygonPointElements as $pointElement) {
-            $content = $pointElement->getContent();
-
-            if (! is_array($content)) {
-                continue;
-            }
-
-            $latElement = $this->firstElement($content, 'pointLatitude');
-            $lonElement = $this->firstElement($content, 'pointLongitude');
-
-            $latText = $this->stringValue($latElement);
-            $lonText = $this->stringValue($lonElement);
-
-            if ($latText !== null && $lonText !== null) {
-                $lat = (float) $this->formatCoordinate($latText);
-                $lon = (float) $this->formatCoordinate($lonText);
-
-                $points[] = [
-                    'lat' => $lat,
-                    'lon' => $lon,
-                ];
-            }
-        }
-
-        return $points;
-    }
-
-    /**
-     * Determine coverage type based on extracted data.
-     *
-     * Priority:
-     * 1. Polygon if polygonPoints exist
-     * 2. Box if all four coordinates (latMin, latMax, lonMin, lonMax) exist
-     * 3. Point if only latMin/lonMin exist
-     *
-     * @param  array<string, mixed>  $coverage
-     */
-    private function determineCoverageType(array $coverage): string
-    {
-        // Check for polygon first (highest priority)
-        if (! empty($coverage['polygonPoints'])) {
-            return 'polygon';
-        }
-
-        // Check for bounding box (all four coordinates)
-        if (($coverage['latMin'] ?? '') !== '' &&
-            ($coverage['latMax'] ?? '') !== '' &&
-            ($coverage['lonMin'] ?? '') !== '' &&
-            ($coverage['lonMax'] ?? '') !== '') {
-            return 'box';
-        }
-
-        // Default to point (includes cases with only latMin/lonMin)
-        return 'point';
-    }
-
-    /**
-     * @param  array<int, array<string, mixed>>  $contributors
-     * @param  array<string, int>  $indexByKey
-     * @param  array<string, mixed>  $contributor
-     * @return array<int, array<string, mixed>>
-     */
-    private function storeContributor(
-        array $contributors,
-        array &$indexByKey,
-        array $contributor,
-    ): array {
-        $keys = $this->buildContributorAggregationKeys($contributor);
-
-        foreach ($keys as $key) {
-            if (! array_key_exists($key, $indexByKey)) {
-                continue;
-            }
-
-            $existingIndex = $indexByKey[$key];
-            $contributors[$existingIndex] = $this->mergeContributorEntries(
-                $contributors[$existingIndex],
-                $contributor,
-            );
-
-            foreach ($keys as $aliasKey) {
-                $indexByKey[$aliasKey] = $existingIndex;
-            }
-
-            return $contributors;
-        }
-
-        $contributors[] = $contributor;
-
-        if (! empty($keys)) {
-            $index = count($contributors) - 1;
-
-            foreach ($keys as $key) {
-                $indexByKey[$key] = $index;
-            }
-        }
-
-        return $contributors;
-    }
-
-    /**
-     * @param  array<string, mixed>  $primary
-     * @param  array<string, mixed>  $incoming
-     * @return array<string, mixed>
-     */
-    private function mergeContributorEntries(array $primary, array $incoming): array
-    {
-        if (($primary['type'] ?? null) !== ($incoming['type'] ?? null)) {
-            return $primary;
-        }
-
-        $primary['roles'] = $this->mergeContributorRoles(
-            is_array($primary['roles'] ?? null) ? $primary['roles'] : [],
-            is_array($incoming['roles'] ?? null) ? $incoming['roles'] : [],
-        );
-
-        if (($primary['type'] ?? null) === 'person') {
-            return $this->mergePersonContributor($primary, $incoming);
-        }
-
-        if (($primary['type'] ?? null) === 'institution') {
-            return $this->mergeInstitutionContributor($primary, $incoming);
-        }
-
-        return $primary;
-    }
-
-    /**
-     * @param  array<int, mixed>  $existing
-     * @param  array<int, mixed>  $incoming
-     * @return array<int, string>
-     */
-    private function mergeContributorRoles(array $existing, array $incoming): array
-    {
-        foreach ($incoming as $role) {
-            if (! is_string($role) || $role === '') {
-                continue;
-            }
-
-            if (! in_array($role, $existing, true)) {
-                $existing[] = $role;
-            }
-        }
-
-        return $existing;
-    }
-
-    /**
-     * @param  array<string, mixed>  $primary
-     * @param  array<string, mixed>  $incoming
-     * @return array<string, mixed>
-     */
-    private function mergePersonContributor(array $primary, array $incoming): array
-    {
-        if (($primary['orcid'] ?? '') === '' && isset($incoming['orcid']) && is_string($incoming['orcid'])) {
-            $primary['orcid'] = $incoming['orcid'];
-        }
-
-        foreach (['firstName', 'lastName'] as $field) {
-            if (
-                (! isset($primary[$field]) || ! is_string($primary[$field]) || trim($primary[$field]) === '')
-                && isset($incoming[$field])
-                && is_string($incoming[$field])
-                && trim($incoming[$field]) !== ''
-            ) {
-                $primary[$field] = $incoming[$field];
-            }
-        }
-
-        $primary['affiliations'] = $this->mergeAffiliations(
-            is_array($primary['affiliations'] ?? null) ? $primary['affiliations'] : [],
-            is_array($incoming['affiliations'] ?? null) ? $incoming['affiliations'] : [],
-        );
-
-        return $primary;
-    }
-
-    /**
-     * @param  array<string, mixed>  $primary
-     * @param  array<string, mixed>  $incoming
-     * @return array<string, mixed>
-     */
-    private function mergeInstitutionContributor(array $primary, array $incoming): array
-    {
-        if (
-            (! isset($primary['institutionName'])
-                || ! is_string($primary['institutionName'])
-                || trim($primary['institutionName']) === '')
-            && isset($incoming['institutionName'])
-            && is_string($incoming['institutionName'])
-            && trim($incoming['institutionName']) !== ''
-        ) {
-            $primary['institutionName'] = $incoming['institutionName'];
-        }
-
-        $primary['affiliations'] = $this->mergeAffiliations(
-            is_array($primary['affiliations'] ?? null) ? $primary['affiliations'] : [],
-            is_array($incoming['affiliations'] ?? null) ? $incoming['affiliations'] : [],
-        );
-
-        return $primary;
-    }
-
-    /**
-     * @param  array<int, mixed>  $existing
-     * @param  array<int, mixed>  $incoming
-     * @return array<int, array{value: string, rorId: ?string}>
-     */
-    private function mergeAffiliations(array $existing, array $incoming): array
-    {
-        $merged = [];
-        $seen = [];
-
-        foreach ([$existing, $incoming] as $group) {
-            foreach ($group as $affiliation) {
-                if (! is_array($affiliation)) {
-                    continue;
-                }
-
-                $value = isset($affiliation['value']) && is_string($affiliation['value'])
-                    ? trim($affiliation['value'])
-                    : '';
-                $rorId = isset($affiliation['rorId']) && is_string($affiliation['rorId'])
-                    ? trim($affiliation['rorId'])
-                    : null;
-
-                if ($rorId === '') {
-                    $rorId = null;
-                }
-
-                if ($value === '' && $rorId === null) {
-                    continue;
-                }
-
-                if ($rorId !== null) {
-                    $key = 'ror:'.Str::lower($rorId);
-                } else {
-                    $key = 'value:'.Str::lower($value);
-                }
-
-                if (isset($seen[$key])) {
-                    $existingIndex = $seen[$key];
-
-                    if ($merged[$existingIndex]['value'] === '' && $value !== '') {
-                        $merged[$existingIndex]['value'] = $value;
-                    }
-
-                    if ($merged[$existingIndex]['rorId'] === null && $rorId !== null) {
-                        $merged[$existingIndex]['rorId'] = $rorId;
-                    }
-
-                    continue;
-                }
-
-                $seen[$key] = count($merged);
-
-                $merged[] = [
-                    'value' => $value,
-                    'rorId' => $rorId,
-                ];
-            }
-        }
-
-        return $merged;
-    }
-
-    /**
-     * @param  array<string, mixed>  $contributor
-     * @return string[]
-     */
-    private function buildContributorAggregationKeys(array $contributor): array
-    {
-        $type = $contributor['type'] ?? null;
-
-        if (! is_string($type)) {
-            return [];
-        }
-
-        if ($type === 'person') {
-            $keys = [];
-            $orcid = $this->normaliseIdentifier($contributor['orcid'] ?? null);
-
-            if ($orcid !== null) {
-                $keys[] = 'person:orcid:'.$orcid;
-            }
-
-            $lastName = $this->normaliseKeyString($contributor['lastName'] ?? null);
-            $firstName = $this->normaliseKeyString($contributor['firstName'] ?? null);
-
-            if ($lastName !== null || $firstName !== null) {
-                $keys[] = 'person:name:'.($lastName ?? '').':'.($firstName ?? '');
-            }
-
-            return $keys;
-        }
-
-        if ($type === 'institution') {
-            $keys = [];
-            $affiliations = is_array($contributor['affiliations'] ?? null)
-                ? $contributor['affiliations']
-                : [];
-
-            foreach ($affiliations as $affiliation) {
-                if (! is_array($affiliation)) {
-                    continue;
-                }
-
-                $rorId = $this->normaliseIdentifier($affiliation['rorId'] ?? null);
-
-                if ($rorId !== null) {
-                    $keys[] = 'institution:ror:'.$rorId;
-                }
-            }
-
-            $institutionName = $this->normaliseKeyString($contributor['institutionName'] ?? null);
-
-            if ($institutionName !== null) {
-                $keys[] = 'institution:name:'.$institutionName;
-            }
-
-            return $keys;
-        }
-
-        return [];
-    }
-
-    private function normaliseIdentifier(mixed $value): ?string
-    {
-        if (! is_string($value)) {
-            return null;
-        }
-
-        $trimmed = trim($value);
-
-        return $trimmed === '' ? null : Str::lower($trimmed);
-    }
-
-    private function normaliseKeyString(mixed $value): ?string
-    {
-        if (! is_string($value)) {
-            return null;
-        }
-
-        $trimmed = trim($value);
-
-        if ($trimmed === '') {
-            return null;
-        }
-
-        $normalised = preg_replace('/\s+/u', ' ', $trimmed);
-
-        if (! is_string($normalised)) {
-            $normalised = $trimmed;
-        }
-
-        return Str::lower($normalised);
-    }
-
-    /**
-     * @return string[]
-     */
-    private function extractContributorRoles(?string $rawRoles): array
-    {
-        if (! is_string($rawRoles)) {
-            return [];
-        }
-
-        $parts = preg_split('/[;,]/', $rawRoles) ?: [];
-
-        if ($parts === []) {
-            $parts = [$rawRoles];
-        }
-
-        $roles = [];
-
-        foreach ($parts as $part) {
-            $trimmed = trim($part);
-
-            if ($trimmed === '') {
-                continue;
-            }
-
-            $resolved = $this->resolveContributorRoleName($trimmed);
-
-            if ($resolved === '') {
-                continue;
-            }
-
-            if (! in_array($resolved, $roles, true)) {
-                $roles[] = $resolved;
-            }
-        }
-
-        return $roles;
-    }
-
-    private function resolveContributorRoleName(string $role): string
-    {
-        $normalisedKey = $this->normaliseContributorRoleKey($role);
-
-        if ($normalisedKey !== null && isset(self::CONTRIBUTOR_ROLE_LABELS[$normalisedKey])) {
-            return self::CONTRIBUTOR_ROLE_LABELS[$normalisedKey];
-        }
-
-        $headline = Str::headline($role);
-
-        return $headline !== '' ? $headline : trim($role);
-    }
-
-    private function normaliseContributorRoleKey(string $role): ?string
-    {
-        $slug = Str::slug($role);
-
-        if ($slug === '') {
-            return null;
-        }
-
-        return str_replace('-', '', $slug);
-    }
-
-    /**
-     * @param  string[]  $roles
-     */
-    private function isInstitutionContributor(?string $nameType, array $roles): bool
-    {
-        if (is_string($nameType)) {
-            $normalised = Str::lower($nameType);
-
-            if ($normalised === 'organizational') {
-                return true;
-            }
-
-            if ($normalised === 'personal') {
-                return false;
-            }
-        }
-
-        return $this->contributorRolesRequireInstitution($roles);
-    }
-
-    /**
-     * @param  string[]  $roles
-     */
-    private function contributorRolesRequireInstitution(array $roles): bool
-    {
-        $hasRoles = false;
-
-        foreach ($roles as $role) {
-            if ($role === '') {
-                continue;
-            }
-
-            $hasRoles = true;
-
-            $key = $this->normaliseContributorRoleKey($role);
-
-            if ($key === null) {
-                return false;
-            }
-
-            if (! in_array($key, self::INSTITUTION_ONLY_CONTRIBUTOR_ROLE_KEYS, true)) {
-                return false;
-            }
-        }
-
-        return $hasRoles;
-    }
-
-    /**
-     * @param  array<string, mixed>  $content
-     * @return array<int, array{value: string, rorId: string|null}>
-     */
-    private function extractInstitutionContributorAffiliations(array $content, string $fallbackLabel): array
-    {
-        $affiliations = $this->extractAffiliations($content);
-
-        if (! empty($affiliations)) {
-            return $affiliations;
-        }
-
-        $identifierElements = $this->allElements($content, 'nameIdentifier');
-
-        foreach ($identifierElements as $element) {
-            $scheme = $element->getAttribute('nameIdentifierScheme');
-            $value = $this->stringValue($element);
-
-            if (! is_string($value) || $value === '') {
-                continue;
-            }
-
-            $resolved = $this->resolveAffiliationByRor(
-                $value,
-                is_string($scheme) ? $scheme : null,
-                $fallbackLabel,
-            );
-
-            if ($resolved !== null) {
-                return [$resolved];
-            }
-        }
-
-        return [];
-    }
-
-    /**
-     * @param  array<string, mixed>  $content
-     * @return array<int, array{value: string, rorId: string|null}>
-     */
-    private function extractAffiliations(array $content): array
-    {
-        $affiliationElements = $this->allElements($content, 'affiliation');
-
-        /** @var array<int, array{value: string, rorId: string|null}> $affiliations */
-        $affiliations = [];
-
-        foreach ($affiliationElements as $element) {
-            $rawValue = $this->stringValue($element);
-
-            $identifier = $element->getAttribute('affiliationIdentifier');
-            $scheme = $element->getAttribute('affiliationIdentifierScheme');
-
-            $resolved = null;
-
-            if (is_string($identifier) && $identifier !== '') {
-                $resolved = $this->resolveAffiliationByRor($identifier, is_string($scheme) ? $scheme : null, $rawValue);
-            }
-
-            if ($resolved !== null) {
-                $affiliations[] = $resolved;
-
-                continue;
-            }
-
-            if (is_string($rawValue) && $rawValue !== '') {
-                $affiliations[] = [
-                    'value' => $rawValue,
-                    'rorId' => null,
-                ];
-            }
-        }
-
-        return $affiliations;
-    }
-
-    /**
-     * @param  array<string, mixed>  $content
-     */
-    private function extractOrcid(array $content): ?string
-    {
-        $identifierElements = $this->allElements($content, 'nameIdentifier');
-
-        foreach ($identifierElements as $element) {
-            $scheme = $element->getAttribute('nameIdentifierScheme');
-
-            if (is_string($scheme) && Str::lower($scheme) !== 'orcid') {
-                continue;
-            }
-
-            $value = $this->stringValue($element);
-
-            if (! is_string($value) || $value === '') {
-                continue;
-            }
-
-            $orcid = $this->canonicaliseOrcid($value);
-
-            if ($orcid !== null) {
-                return $orcid;
-            }
-        }
-
-        return null;
-    }
-
-    private function canonicaliseOrcid(string $value): ?string
-    {
-        $trimmed = trim($value);
-
-        if ($trimmed === '') {
-            return null;
-        }
-
-        if (preg_match('#^https?://orcid\.org/#i', $trimmed) === 1) {
-            // Use PHP 8.5's RFC 3986 compliant URI parser
-            $path = UriHelper::getPath($trimmed);
-            $identifier = is_string($path) ? trim($path, '/') : '';
-        } else {
-            $identifier = trim($trimmed, '/');
-        }
-
-        if ($identifier === '') {
-            return null;
-        }
-
-        // Return only the ORCID identifier (e.g., "0000-0001-5727-2427")
-        // without the URL prefix, as the frontend expects this format
-        return $identifier;
-    }
-
-    /**
-     * Extract MSL Laboratory ID (labid) from contributor content
-     *
-     * @param  array<string, mixed>  $content
-     */
-    private function extractLabId(array $content): ?string
-    {
-        $identifierElements = $this->allElements($content, 'nameIdentifier');
-
-        foreach ($identifierElements as $element) {
-            $scheme = $element->getAttribute('nameIdentifierScheme');
-
-            if (is_string($scheme) && Str::lower($scheme) === 'labid') {
-                $value = $this->stringValue($element);
-
-                if (is_string($value) && trim($value) !== '') {
-                    return trim($value);
-                }
-            }
-        }
-
-        return null;
-    }
-
-    /**
-     * Extract and enrich MSL Laboratory data
-     *
-     * @param  array<string, mixed>  $content
-     * @return array{identifier: string, name: string, affiliation_name: string, affiliation_ror: string}|null
-     */
-    private function extractMslLaboratory(array $content, string $labId): ?array
-    {
-        $mslService = app(MslLaboratoryService::class);
-
-        // Get laboratory name from XML
-        $nameElement = $this->firstElement($content, 'contributorName');
-        $labName = $this->stringValue($nameElement);
-
-        // Get affiliation info from XML
-        $affiliationElement = $this->firstElement($content, 'affiliation');
-        $affiliationName = $this->stringValue($affiliationElement);
-
-        // Extract ROR if present - check scheme first, then fallback to identifier inspection
-        $affiliationRor = null;
-        $affiliationIdentifier = $affiliationElement?->getAttribute('affiliationIdentifier');
-        $affiliationScheme = $affiliationElement?->getAttribute('affiliationIdentifierScheme');
-
-        Log::debug('Extracting MSL Lab affiliation', [
-            'labId' => $labId,
-            'affiliationName' => $affiliationName,
-            'affiliationIdentifier' => $affiliationIdentifier,
-            'affiliationScheme' => $affiliationScheme,
-        ]);
-
-        if ($affiliationIdentifier) {
-            // Explicit ROR scheme OR identifier looks like a ROR
-            $isRor = ($affiliationScheme === 'ROR') ||
-                     str_contains(strtolower($affiliationIdentifier), 'ror.org');
-
-            if ($isRor) {
-                // Normalize to full URL if not already
-                if (str_starts_with($affiliationIdentifier, 'http')) {
-                    $affiliationRor = $affiliationIdentifier;
-                } else {
-                    $affiliationRor = 'https://ror.org/'.$affiliationIdentifier;
-                }
-
-                Log::debug('ROR identified and normalized', [
-                    'original' => $affiliationIdentifier,
-                    'normalized' => $affiliationRor,
-                ]);
-            }
-        }
-
-        // Enrich with data from Utrecht vocabulary (using labid)
-        $enrichedLab = $mslService->enrichLaboratoryData(
-            $labId,
-            $labName,
-            $affiliationName,
-            $affiliationRor
-        );
-
-        // Validate that we have at least a name
-        if (empty($enrichedLab['name'])) {
-            Log::warning('MSL Laboratory extracted without name', [
-                'labId' => $labId,
-            ]);
-
-            return null;
-        }
-
-        return $enrichedLab;
-    }
-
-    /**
-     * @return array{value: string, rorId: string}|null
-     */
-    private function resolveAffiliationByRor(string $identifier, ?string $scheme, ?string $fallback): ?array
-    {
-        if (! $this->isRorIdentifier($identifier, $scheme)) {
-            return null;
-        }
-
-        return $this->rorLookupService->resolveWithFallback($identifier, $fallback);
-    }
-
-    private function isRorIdentifier(string $identifier, ?string $scheme): bool
-    {
-        // Only treat as ROR if scheme explicitly indicates ROR
-        if (is_string($scheme) && Str::lower($scheme) === 'ror') {
-            return true;
-        }
-
-        // Also accept if identifier contains ror.org (URL format)
-        return Str::contains(Str::lower($identifier), 'ror.org');
-    }
-
-    /**
-     * @param  array<string, mixed>  $content
-     */
-    private function firstElement(array $content, string $key): ?Element
-    {
-        return XmlElementHelpers::firstElementByKey($content, $key);
-    }
-
-    /**
-     * @param  array<string, mixed>  $content
-     * @return Element[]
-     */
-    private function allElements(array $content, string $key): array
-    {
-        return XmlElementHelpers::allElementsByKey($content, $key);
-    }
-
-    private function stringValue(?Element $element): ?string
-    {
-        return XmlElementHelpers::stringValue($element);
-    }
-
-    /**
-     * @return array{givenName: string|null, familyName: string|null}
-     */
-    private function splitCreatorName(?string $name): array
-    {
-        return XmlElementHelpers::splitCreatorName($name);
-    }
-
-    /**
-     * Extract GCMD keywords from the XML.
-     *
-     * @return array<int, array{uuid: string, id: string, text: string, path: string, scheme: string, schemeURI?: string, classificationCode?: string}>
-     */
-    private function extractGcmdKeywords(XmlReader $reader): array
-    {
-        $subjectElements = $reader
-            ->xpathElement('//*[local-name()="subjects"]/*[local-name()="subject"]')
-            ->get();
-
-        $keywords = [];
-
-        foreach ($subjectElements as $element) {
-            $scheme = trim((string) $element->getAttribute('subjectScheme'));
-            $valueUri = trim((string) $element->getAttribute('valueURI'));
-            $classificationCode = trim((string) $element->getAttribute('classificationCode'));
-            $content = $this->stringValue($element);
-
-            if (! $scheme || ! $content) {
-                continue;
-            }
-
-            // Only process GCMD keywords (Science Keywords, Platforms, Instruments)
-            $isGcmdKeyword = stripos($scheme, 'Science Keywords') !== false ||
-                            stripos($scheme, 'Platforms') !== false ||
-                            stripos($scheme, 'Instruments') !== false;
-
-            if (! $isGcmdKeyword) {
-                // Handle non-GCMD schemes that have a classificationCode or valueURI
-                // (e.g., International Chronostratigraphic Chart, custom vocabularies)
-                if ($valueUri !== '' || $classificationCode !== '') {
-                    $schemeUri = trim((string) $element->getAttribute('schemeURI'));
-
-                    // Parse leaf text from hierarchical path (e.g., "Phanerozoic > Mesozoic > Jurassic" → "Jurassic")
-                    $fullPath = trim($content);
-                    $pathParts = array_map('trim', explode('>', $fullPath));
-                    $leafText = end($pathParts) ?: $fullPath;
-
-                    $keyword = [
-                        'uuid' => '',
-                        'id' => $valueUri !== '' ? $valueUri : $classificationCode,
-                        'text' => $leafText,
-                        'path' => $fullPath,
-                        'scheme' => $scheme,
-                    ];
-
-                    if ($schemeUri !== '') {
-                        $keyword['schemeURI'] = $schemeUri;
-                    }
-
-                    if ($classificationCode !== '') {
-                        $keyword['classificationCode'] = $classificationCode;
-                    }
-
-                    $keywords[] = $keyword;
-                }
-
-                continue;
-            }
-
-            if (! $valueUri) {
-                continue;
-            }
-
-            // Extract UUID from valueURI using shared utility
-            $uuid = GcmdUriHelper::extractUuid($valueUri);
-
-            if (! $uuid) {
-                continue;
-            }
-
-            // Build full ID URL using shared utility
-            $id = GcmdUriHelper::buildConceptUri($uuid);
-
-            // Parse the hierarchical path using shared utility
-            // Example: "Science Keywords > EARTH SCIENCE > ATMOSPHERE" -> ["EARTH SCIENCE", "ATMOSPHERE"]
-            $pathArray = XmlKeywordExtractor::parseGcmdPath($content);
-
-            // Convert path array to string format expected by frontend/validation
-            // Example: ["EARTH SCIENCE", "ATMOSPHERE"] -> "EARTH SCIENCE > ATMOSPHERE"
-            $pathString = implode(' > ', $pathArray);
-
-            // Extract the text (last element of the path, which is the keyword itself)
-            // If path is empty, use the content directly
-            $text = array_last($pathArray) ?? $content;
-
-            // Normalize scheme name (e.g., "NASA/GCMD Earth Science Keywords" -> "Science Keywords")
-            $normalizedScheme = $scheme;
-            if (stripos($scheme, 'Science') !== false) {
-                $normalizedScheme = 'Science Keywords';
-            } elseif (stripos($scheme, 'Platform') !== false) {
-                $normalizedScheme = 'Platforms';
-            } elseif (stripos($scheme, 'Instrument') !== false) {
-                $normalizedScheme = 'Instruments';
-            }
-
-            $keyword = [
-                'uuid' => $uuid,
-                'id' => $id,
-                'text' => $text,
-                'path' => $pathString,
-                'scheme' => $normalizedScheme,
-            ];
-
-            if ($classificationCode !== '') {
-                $keyword['classificationCode'] = $classificationCode;
-            }
-
-            $keywords[] = $keyword;
-        }
-
-        return $keywords;
-    }
-
-    /**
-     * Extract funding references from DataCite XML.
-     *
-     * @return array<int, array{funderName: string, funderIdentifier: string|null, funderIdentifierType: string|null, awardNumber: string|null, awardUri: string|null, awardTitle: string|null}>
-     */
-    private function extractFundingReferences(XmlReader $reader): array
-    {
-        $fundingElements = $reader
-            ->xpathElement('//*[local-name()="fundingReferences"]/*[local-name()="fundingReference"]')
-            ->get();
-
-        $fundingReferences = [];
-
-        foreach ($fundingElements as $fundingElement) {
-            $content = $fundingElement->getContent();
-
-            if (! is_array($content)) {
-                continue;
-            }
-
-            // Extract funderName (required in DataCite schema)
-            $funderNameElement = $this->firstElement($content, 'funderName');
-            $funderName = $this->stringValue($funderNameElement);
-
-            if (! $funderName) {
-                // Skip if funderName is missing (required field)
-                continue;
-            }
-
-            // Extract funderIdentifier and its type
-            $funderIdentifierElement = $this->firstElement($content, 'funderIdentifier');
-            $funderIdentifier = $this->stringValue($funderIdentifierElement);
-            $funderIdentifierType = $funderIdentifierElement?->getAttribute('funderIdentifierType');
-
-            // Extract awardNumber and awardURI
-            $awardNumberElement = $this->firstElement($content, 'awardNumber');
-            $awardNumber = $this->stringValue($awardNumberElement);
-            $awardUri = $awardNumberElement?->getAttribute('awardURI');
-
-            // Extract awardTitle
-            $awardTitleElement = $this->firstElement($content, 'awardTitle');
-            $awardTitle = $this->stringValue($awardTitleElement);
-
-            $fundingReferences[] = [
-                'funderName' => $funderName,
-                'funderIdentifier' => $funderIdentifier,
-                'funderIdentifierType' => $funderIdentifierType,
-                'awardNumber' => $awardNumber,
-                'awardUri' => $awardUri,
-                'awardTitle' => $awardTitle,
-            ];
-        }
-
-        return $fundingReferences;
-    }
-
-    /**
-     * Extract contact information (email, website) from ISO 19115 pointOfContact elements.
-     *
-     * Parses the MD_Metadata section of the XML to find pointOfContact entries
-     * with individualName and their associated email addresses and websites.
-     *
-     * @return array<string, array{email: string, website: string}>
-     *         Key is normalized name "familyname, givenname" (lowercase, trimmed)
-     */
-    private function extractContactInfoFromIso(XmlReader $reader): array
-    {
-        $contactInfo = [];
-
-        // Query all pointOfContact elements within MD_DataIdentification
-        // These are in the ISO 19115 namespace (gmd)
-        $pointOfContactElements = $reader
-            ->xpathElement('//*[local-name()="identificationInfo"]//*[local-name()="pointOfContact"]/*[local-name()="CI_ResponsibleParty"]')
-            ->get();
-
-        foreach ($pointOfContactElements as $element) {
-            $content = $element->getContent();
-
-            if (! is_array($content)) {
-                continue;
-            }
-
-            // Extract individualName (format: "Familyname, Givenname")
-            $individualName = $this->extractIsoCharacterString($content, 'individualName');
-
-            if ($individualName === null || $individualName === '') {
-                continue;
-            }
-
-            $normalizedName = $this->normalizeNameForMatching($individualName);
-
-            // Skip if we already have contact info for this person (use first found)
-            if (isset($contactInfo[$normalizedName])) {
-                continue;
-            }
-
-            // Extract email from nested CI_Contact/address/CI_Address/electronicMailAddress
-            $email = $this->extractIsoEmail($content);
-
-            // Extract website from nested CI_Contact/onlineResource/CI_OnlineResource/linkage/URL
-            $website = $this->extractIsoWebsite($content);
-
-            $contactInfo[$normalizedName] = [
-                'email' => $email ?? '',
-                'website' => $website ?? '',
-            ];
-        }
-
-        return $contactInfo;
-    }
-
-    /**
-     * Get a direct child element from ISO 19115 content without normalization.
-     *
-     * Unlike firstElement(), this method does not flatten nested structures.
-     * This is necessary for ISO 19115 parsing where we need to traverse
-     * the deep hierarchy step by step.
-     *
-     * @param  array<string, mixed>  $content
-     */
-    private function getIsoChildElement(array $content, string $key): ?Element
-    {
-        if (! array_key_exists($key, $content)) {
-            return null;
-        }
-
-        $value = $content[$key];
-
-        return $value instanceof Element ? $value : null;
-    }
-
-    /**
-     * Extract a CharacterString value from ISO 19115 content.
-     *
-     * ISO 19115 wraps most text values in gco:CharacterString elements.
-     *
-     * @param  array<string, mixed>  $content
-     */
-    private function extractIsoCharacterString(array $content, string $elementName): ?string
-    {
-        $element = $this->getIsoChildElement($content, $elementName);
-
-        if ($element === null) {
-            return null;
-        }
-
-        $innerContent = $element->getContent();
-
-        // Check if it's directly a string
-        if (is_string($innerContent)) {
-            return trim($innerContent);
-        }
-
-        // Check for nested CharacterString (gco:CharacterString with or without prefix)
-        if (is_array($innerContent)) {
-            // Try with namespace prefix first, then without
-            $charString = $this->getIsoChildElement($innerContent, 'gco:CharacterString')
-                ?? $this->getIsoChildElement($innerContent, 'CharacterString');
-
-            if ($charString !== null) {
-                $value = $charString->getContent();
-                if (is_string($value)) {
-                    return trim($value);
-                }
-            }
-        }
-
-        return null;
-    }
-
-    /**
-     * Extract email address from ISO 19115 CI_ResponsibleParty content.
-     *
-     * Path: contactInfo/CI_Contact/address/CI_Address/electronicMailAddress/CharacterString
-     *
-     * @param  array<string, mixed>  $content
-     */
-    private function extractIsoEmail(array $content): ?string
-    {
-        $contactInfoElement = $this->getIsoChildElement($content, 'contactInfo');
-        if ($contactInfoElement === null) {
-            return null;
-        }
-
-        $contactInfoContent = $contactInfoElement->getContent();
-        if (! is_array($contactInfoContent)) {
-            return null;
-        }
-
-        $ciContact = $this->getIsoChildElement($contactInfoContent, 'CI_Contact');
-        if ($ciContact === null) {
-            return null;
-        }
-
-        $ciContactContent = $ciContact->getContent();
-        if (! is_array($ciContactContent)) {
-            return null;
-        }
-
-        $address = $this->getIsoChildElement($ciContactContent, 'address');
-        if ($address === null) {
-            return null;
-        }
-
-        $addressContent = $address->getContent();
-        if (! is_array($addressContent)) {
-            return null;
-        }
-
-        $ciAddress = $this->getIsoChildElement($addressContent, 'CI_Address');
-        if ($ciAddress === null) {
-            return null;
-        }
-
-        $ciAddressContent = $ciAddress->getContent();
-        if (! is_array($ciAddressContent)) {
-            return null;
-        }
-
-        return $this->extractIsoCharacterString($ciAddressContent, 'electronicMailAddress');
-    }
-
-    /**
-     * Extract website URL from ISO 19115 CI_ResponsibleParty content.
-     *
-     * Path: contactInfo/CI_Contact/onlineResource/CI_OnlineResource/linkage/URL
-     *
-     * @param  array<string, mixed>  $content
-     */
-    private function extractIsoWebsite(array $content): ?string
-    {
-        $contactInfoElement = $this->getIsoChildElement($content, 'contactInfo');
-        if ($contactInfoElement === null) {
-            return null;
-        }
-
-        $contactInfoContent = $contactInfoElement->getContent();
-        if (! is_array($contactInfoContent)) {
-            return null;
-        }
-
-        $ciContact = $this->getIsoChildElement($contactInfoContent, 'CI_Contact');
-        if ($ciContact === null) {
-            return null;
-        }
-
-        $ciContactContent = $ciContact->getContent();
-        if (! is_array($ciContactContent)) {
-            return null;
-        }
-
-        $onlineResource = $this->getIsoChildElement($ciContactContent, 'onlineResource');
-        if ($onlineResource === null) {
-            return null;
-        }
-
-        $onlineResourceContent = $onlineResource->getContent();
-        if (! is_array($onlineResourceContent)) {
-            return null;
-        }
-
-        $ciOnlineResource = $this->getIsoChildElement($onlineResourceContent, 'CI_OnlineResource');
-        if ($ciOnlineResource === null) {
-            return null;
-        }
-
-        $ciOnlineResourceContent = $ciOnlineResource->getContent();
-        if (! is_array($ciOnlineResourceContent)) {
-            return null;
-        }
-
-        $linkage = $this->getIsoChildElement($ciOnlineResourceContent, 'linkage');
-        if ($linkage === null) {
-            return null;
-        }
-
-        $linkageContent = $linkage->getContent();
-        if (! is_array($linkageContent)) {
-            return null;
-        }
-
-        $urlElement = $this->getIsoChildElement($linkageContent, 'URL');
-        if ($urlElement === null) {
-            return null;
-        }
-
-        $url = $urlElement->getContent();
-
-        return is_string($url) ? trim($url) : null;
-    }
-
-    /**
-     * Normalize a name for matching purposes.
-     *
-     * Handles both "Familyname, Givenname" format and separate first/last name.
-     * Returns lowercase, trimmed version for case-insensitive comparison.
-     *
-     * @param  string  $name  Name in "Familyname, Givenname" format
-     */
-    private function normalizeNameForMatching(string $name): string
-    {
-        return mb_strtolower(trim($name));
-    }
-
-    /**
-     * Build a normalized name key from first and last name.
-     *
-     * @return string Name in "familyname, givenname" format (lowercase)
-     */
-    private function buildNameKey(string $familyName, string $firstName): string
-    {
-        return mb_strtolower(trim($familyName).', '.trim($firstName));
-    }
-
-    /**
-     * Merge ContactPerson contributors into authors array.
-     *
-     * Matching priority:
-     * 1. ORCID (if both have it)
-     * 2. Name (familyName + givenName, case-insensitive)
-     *
-     * If match found: Set isContact=true, add email/website from ISO
-     * If no match: Add as new author with isContact=true
-     *
-     * @param  array<int, array<string, mixed>>  $authors  Extracted authors from creators
-     * @param  array<int, array<string, mixed>>  $contactPersons  ContactPerson contributors
-     * @param  array<string, array{email: string, website: string}>  $isoContactInfo  Email/website map from ISO
-     * @return array<int, array<string, mixed>>  Modified authors array
-     */
-    private function mergeContactPersonsIntoAuthors(array $authors, array $contactPersons, array $isoContactInfo): array
-    {
-        foreach ($contactPersons as $cp) {
-            // Skip institution-type contact persons (not supported in ERNIE)
-            if (($cp['type'] ?? '') === 'institution') {
-                continue;
-            }
-
-            $matched = false;
-
-            // 1. Try ORCID match first (most reliable)
-            $cpOrcid = $cp['orcid'] ?? '';
-            if ($cpOrcid !== '') {
-                foreach ($authors as &$author) {
-                    if (($author['type'] ?? '') === 'person' && ($author['orcid'] ?? '') === $cpOrcid) {
-                        $author['isContact'] = true;
-                        $this->enrichAuthorWithContactInfo($author, $isoContactInfo);
-                        $matched = true;
-                        break;
-                    }
-                }
-                unset($author);
-            }
-
-            // 2. Fallback to name match
-            if (! $matched) {
-                $cpLastName = trim($cp['lastName'] ?? '');
-                $cpFirstName = trim($cp['firstName'] ?? '');
-
-                // Skip name matching if lastName is empty (firstName alone is not reliable enough)
-                if ($cpLastName === '') {
-                    continue;
-                }
-
-                $cpNameKey = $this->buildNameKey($cpLastName, $cpFirstName);
-
-                foreach ($authors as &$author) {
-                    if (($author['type'] ?? '') === 'person') {
-                        $authorNameKey = $this->buildNameKey($author['lastName'] ?? '', $author['firstName'] ?? '');
-                        if ($cpNameKey === $authorNameKey) {
-                            $author['isContact'] = true;
-                            $this->enrichAuthorWithContactInfo($author, $isoContactInfo);
-                            $matched = true;
-                            break;
-                        }
-                    }
-                }
-                unset($author);
-            }
-
-            // 3. No match found - add as new author
-            if (! $matched) {
-                $newAuthor = [
-                    'type' => 'person',
-                    'orcid' => $cp['orcid'] ?? '',
-                    'firstName' => $cp['firstName'] ?? '',
-                    'lastName' => $cp['lastName'] ?? '',
-                    'affiliations' => $cp['affiliations'] ?? [],
-                    'isContact' => true,
-                    'email' => '',
-                    'website' => '',
-                ];
-                $this->enrichAuthorWithContactInfo($newAuthor, $isoContactInfo);
-                $authors[] = $newAuthor;
-            }
-        }
-
-        return $authors;
-    }
-
-    /**
-     * Enrich an author entry with email and website from ISO contact info.
-     *
-     * @param  array<string, mixed>  $author  Author entry (modified by reference)
-     * @param  array<string, array{email: string, website: string}>  $isoContactInfo  Contact info map
-     */
-    private function enrichAuthorWithContactInfo(array &$author, array $isoContactInfo): void
-    {
-        $nameKey = $this->buildNameKey($author['lastName'] ?? '', $author['firstName'] ?? '');
-
-        if (isset($isoContactInfo[$nameKey])) {
-            $author['email'] = $isoContactInfo[$nameKey]['email'];
-            $author['website'] = $isoContactInfo[$nameKey]['website'];
-        } else {
-            // Ensure fields exist even if no match found
-            $author['email'] = $author['email'] ?? '';
-            $author['website'] = $author['website'] ?? '';
-        }
-    }
-
-    /**
-     * Parse temporal coverage components from a Coverage-type date entry.
-     *
-     * Extracts full datetime components (date, time, timezone) from the raw
-     * date value to preserve time information that normalizeDateString() strips.
-     *
-     * @param  array<string, string>|null  $temporalCoverage  Coverage date entry with rawValue
-     * @return array{startDate: string, endDate: string, startTime: string, endTime: string, timezone: string}
-     */
-    private function parseTemporalCoverageComponents(?array $temporalCoverage): array
-    {
-        $default = [
-            'startDate' => '',
-            'endDate' => '',
-            'startTime' => '',
-            'endTime' => '',
-            'timezone' => 'UTC',
-        ];
-
-        if ($temporalCoverage === null) {
-            return $default;
-        }
-
-        $rawValue = $temporalCoverage['rawValue'] ?? '';
-
-        if ($rawValue === '') {
-            return [
-                'startDate' => $temporalCoverage['startDate'] ?? '',
-                'endDate' => $temporalCoverage['endDate'] ?? '',
-                'startTime' => '',
-                'endTime' => '',
-                'timezone' => 'UTC',
-            ];
-        }
-
-        if (str_contains($rawValue, '/')) {
-            [$startRaw, $endRaw] = explode('/', $rawValue, 2);
-            $startParts = $this->parseDateTimeComponents(trim($startRaw));
-            $endParts = $this->parseDateTimeComponents(trim($endRaw));
-        } else {
-            $startParts = $this->parseDateTimeComponents($rawValue);
-            $endParts = ['date' => '', 'time' => '', 'timezone' => ''];
-        }
-
-        // Determine timezone: use explicit timezone from start or end, otherwise
-        // only default to 'UTC' if no time component exists (pure date values)
-        $hasTimeComponent = $startParts['time'] !== '' || $endParts['time'] !== '';
-        $resolvedTimezone = $startParts['timezone'] ?: $endParts['timezone'];
-
-        if ($resolvedTimezone === '' && ! $hasTimeComponent) {
-            $resolvedTimezone = 'UTC';
-        }
-
-        return [
-            'startDate' => $startParts['date'],
-            'endDate' => $endParts['date'],
-            'startTime' => $startParts['time'],
-            'endTime' => $endParts['time'],
-            'timezone' => $resolvedTimezone,
-        ];
-    }
-
-    /**
-     * Parse a date/datetime string into separate date, time, and timezone components.
-     *
-     * @return array{date: string, time: string, timezone: string}
-     */
-    private function parseDateTimeComponents(string $value): array
-    {
-        $value = trim($value);
-
-        if ($value === '') {
-            return ['date' => '', 'time' => '', 'timezone' => ''];
-        }
-
-        // Only attempt full datetime parsing if the value contains a time separator
-        if (str_contains($value, 'T')) {
-            try {
-                // Detect whether the original string contains an explicit timezone designator
-                // (Z, +HH:MM, -HH:MM) to avoid environment-dependent parsing via php.ini date.timezone
-                $timePart = substr($value, (int) strpos($value, 'T') + 1);
-                $hasExplicitTimezone = (bool) preg_match('/[Zz]$|[+-]\d{2}:\d{2}$/', $timePart);
-
-                $dt = $hasExplicitTimezone
-                    ? new \DateTimeImmutable($value)
-                    : new \DateTimeImmutable($value, new \DateTimeZone('UTC'));
-
-                $date = $dt->format('Y-m-d');
-                $time = (int) $dt->format('s') !== 0 ? $dt->format('H:i:s') : $dt->format('H:i');
-                $timezone = $hasExplicitTimezone ? $this->resolveTimezone($dt) : '';
-
-                return ['date' => $date, 'time' => $time, 'timezone' => $timezone];
-            } catch (\Exception) {
-                // Fall through to simple date normalization
-            }
-        }
-
-        // For plain dates without time, use existing normalization
-        return ['date' => $this->normalizeDateString($value), 'time' => '', 'timezone' => ''];
-    }
-
-    /**
-     * Extract the timezone identifier from a DateTimeImmutable.
-     *
-     * Returns the timezone name as-is (IANA name or UTC offset string like "+02:00").
-     * Numeric offsets are kept verbatim because mapping them to IANA zones is
-     * inherently ambiguous and non-deterministic across PHP versions.
-     * "+00:00" is normalized to "UTC" since they are semantically identical.
-     */
-    private function resolveTimezone(\DateTimeImmutable $dt): string
-    {
-        $tz = $dt->getTimezone();
-        $tzName = $tz->getName();
-
-        if ($tzName === '' || $tzName === '+00:00' || $tzName === 'Z') {
-            return 'UTC';
-        }
-
-        return $tzName;
     }
 
     /**
